refactor: [client] Modal 공동컴포넌트로 관리
